Make ChannelOutboundBuffer an interface and so allow to write Channel implementations that not extend AbstractChannel
Motivation:

ChannelOutboundBuffer was a final class that had a package-private constructor that took AbstractChannel as argument. Due of this it was impossible to write a Channel implementation that not extend AbstractChannel.

Modifications:

- Make ChannelOutboundBuffer an interface
- Add DefaultChannelOutboundBuffer implementation that is used by AbstractChannel
- Factor out code that produced ByteBuffer to NioSocketChannel as it is the only place we use it
- Add unit tests

Result:

API cleanup
