+++ transport/src/main/java/io/netty5/channel/DefaultChannelOutboundBuffer.java
@@ -1,0 +1,631 @@
+/*
+ * Copyright 2013 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package io.netty5.channel;
+
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufConvertible;
+import io.netty.buffer.ByteBufHolder;
+import io.netty.buffer.Unpooled;
+import io.netty5.buffer.api.Buffer;
+import io.netty5.util.concurrent.Promise;
+import io.netty5.util.internal.ObjectPool;
+import io.netty5.util.internal.ObjectPool.Handle;
+import io.netty5.util.internal.PromiseNotificationUtil;
+import io.netty5.util.internal.SilentDispose;
+import io.netty5.util.internal.SystemPropertyUtil;
+import io.netty5.util.internal.logging.InternalLogger;
+import io.netty5.util.internal.logging.InternalLoggerFactory;
+
+import java.nio.ByteBuffer;
+import java.nio.channels.ClosedChannelException;
+import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
+import java.util.concurrent.atomic.AtomicLongFieldUpdater;
+
+import static java.util.Objects.requireNonNull;
+
+/**
+ * (Transport implementors only) an internal data structure used by {@link AbstractChannel} to store its pending
+ * outbound write requests.
+ * <p>
+ * All methods must be called by a transport implementation from an I/O thread, except the following ones:
+ * <ul>
+ * <li>{@link #size()} and {@link #isEmpty()}</li>
+ * <li>{@link #isWritable()}</li>
+ * <li>{@link #getUserDefinedWritability(int)} and {@link #setUserDefinedWritability(int, boolean)}</li>
+ * </ul>
+ * </p>
+ */
+final class DefaultChannelOutboundBuffer implements ChannelOutboundBuffer {
+    // Assuming a 64-bit JVM:
+    //  - 16 bytes object header
+    //  - 6 reference fields
+    //  - 2 long fields
+    //  - 2 int fields
+    //  - 1 boolean field
+    //  - padding
+    static final int CHANNEL_OUTBOUND_BUFFER_ENTRY_OVERHEAD =
+            SystemPropertyUtil.getInt("io.netty5.transport.outboundBufferEntrySizeOverhead", 96);
+
+    private static final InternalLogger logger = InternalLoggerFactory.getInstance(DefaultChannelOutboundBuffer.class);
+
+    private final Channel channel;
+
+    // Entry(flushedEntry) --> ... Entry(unflushedEntry) --> ... Entry(tailEntry)
+    //
+    // The Entry that is the first in the linked-list structure that was flushed
+    private Entry flushedEntry;
+    // The Entry which is the first unflushed in the linked-list structure
+    private Entry unflushedEntry;
+    // The Entry which represents the tail of the buffer
+    private Entry tailEntry;
+    // The number of flushed entries that are not written yet
+    private int flushed;
+
+    private boolean inFail;
+
+    private static final AtomicLongFieldUpdater<DefaultChannelOutboundBuffer> TOTAL_PENDING_SIZE_UPDATER =
+            AtomicLongFieldUpdater.newUpdater(DefaultChannelOutboundBuffer.class, "totalPendingSize");
+
+    @SuppressWarnings("UnusedDeclaration")
+    private volatile long totalPendingSize;
+
+    private static final AtomicIntegerFieldUpdater<DefaultChannelOutboundBuffer> UNWRITABLE_UPDATER =
+            AtomicIntegerFieldUpdater.newUpdater(DefaultChannelOutboundBuffer.class, "unwritable");
+
+    @SuppressWarnings("UnusedDeclaration")
+    private volatile int unwritable;
+
+    private volatile Runnable fireChannelWritabilityChangedTask;
+
+    DefaultChannelOutboundBuffer(Channel channel) {
+        this.channel = channel;
+    }
+
+    @Override
+    public void addMessage(Object msg, int size, Promise<Void> promise) {
+        Entry entry = Entry.newInstance(msg, size, total(msg), promise);
+        if (tailEntry == null) {
+            flushedEntry = null;
+        } else {
+            Entry tail = tailEntry;
+            tail.next = entry;
+        }
+        tailEntry = entry;
+        if (unflushedEntry == null) {
+            unflushedEntry = entry;
+        }
+
+        // increment pending bytes after adding message to the unflushed arrays.
+        // See https://github.com/netty/netty/issues/1619
+        incrementPendingOutboundBytes(entry.pendingSize, false);
+    }
+
+    @Override
+    public void addFlush() {
+        // There is no need to process all entries if there was already a flush before and no new messages
+        // where added in the meantime.
+        //
+        // See https://github.com/netty/netty/issues/2577
+        Entry entry = unflushedEntry;
+        if (entry != null) {
+            if (flushedEntry == null) {
+                // there is no flushedEntry yet, so start with the entry
+                flushedEntry = entry;
+            }
+            do {
+                flushed ++;
+                if (!entry.promise.setUncancellable()) {
+                    // Was cancelled so make sure we free up memory and notify about the freed bytes
+                    int pending = entry.cancel();
+                    decrementPendingOutboundBytes(pending, false, true);
+                }
+                entry = entry.next;
+            } while (entry != null);
+
+            // All flushed so reset unflushedEntry
+            unflushedEntry = null;
+        }
+    }
+
+    /**
+     * Increment the pending bytes which will be written at some point.
+     * This method is thread-safe!
+     */
+    void incrementPendingOutboundBytes(long size) {
+        incrementPendingOutboundBytes(size, true);
+    }
+
+    private void incrementPendingOutboundBytes(long size, boolean invokeLater) {
+        if (size == 0) {
+            return;
+        }
+
+        long newWriteBufferSize = TOTAL_PENDING_SIZE_UPDATER.addAndGet(this, size);
+        if (newWriteBufferSize > channel.config().getWriteBufferHighWaterMark()) {
+            setUnwritable(invokeLater);
+        }
+    }
+
+    /**
+     * Decrement the pending bytes which will be written at some point.
+     * This method is thread-safe!
+     */
+    void decrementPendingOutboundBytes(long size) {
+        decrementPendingOutboundBytes(size, true, true);
+    }
+
+    private void decrementPendingOutboundBytes(long size, boolean invokeLater, boolean notifyWritability) {
+        if (size == 0) {
+            return;
+        }
+
+        long newWriteBufferSize = TOTAL_PENDING_SIZE_UPDATER.addAndGet(this, -size);
+        if (notifyWritability && newWriteBufferSize < channel.config().getWriteBufferLowWaterMark()) {
+            setWritable(invokeLater);
+        }
+    }
+
+    private static long total(Object msg) {
+        if (msg instanceof Buffer) {
+            return ((Buffer) msg).readableBytes();
+        }
+        if (msg instanceof FileRegion) {
+            return ((FileRegion) msg).count();
+        }
+        if (msg instanceof ByteBufHolder) {
+            return ((ByteBufHolder) msg).content().readableBytes();
+        }
+        if (msg instanceof ByteBufConvertible) {
+            return ((ByteBufConvertible) msg).asByteBuf().readableBytes();
+        }
+        return -1;
+    }
+
+    @Override
+    public Object current() {
+        Entry entry = flushedEntry;
+        if (entry == null) {
+            return null;
+        }
+
+        return entry.msg;
+    }
+
+    @Override
+    public long currentProgress() {
+        Entry entry = flushedEntry;
+        if (entry == null) {
+            return 0;
+        }
+        return entry.progress;
+    }
+
+    @Override
+    public void progress(long amount) {
+        Entry e = flushedEntry;
+        assert e != null;
+        e.progress += amount;
+    }
+
+    @Override
+    public boolean remove() {
+        Entry e = flushedEntry;
+        if (e == null) {
+            return false;
+        }
+        Object msg = e.msg;
+
+        Promise<Void> promise = e.promise;
+        int size = e.pendingSize;
+
+        removeEntry(e);
+
+        if (!e.cancelled) {
+            // only release message, notify and decrement if it was not canceled before.
+            SilentDispose.trySilentDispose(msg, logger);
+            safeSuccess(promise);
+            decrementPendingOutboundBytes(size, false, true);
+        }
+
+        // recycle the entry
+        e.recycle();
+
+        return true;
+    }
+
+    @Override
+    public boolean remove(Throwable cause) {
+        return remove0(cause, true);
+    }
+
+    private boolean remove0(Throwable cause, boolean notifyWritability) {
+        Entry e = flushedEntry;
+        if (e == null) {
+            return false;
+        }
+        Object msg = e.msg;
+
+        Promise<Void> promise = e.promise;
+        int size = e.pendingSize;
+
+        removeEntry(e);
+
+        if (!e.cancelled) {
+            // only release message, fail and decrement if it was not canceled before.
+            SilentDispose.trySilentDispose(msg, logger);
+
+            safeFail(promise, cause);
+            decrementPendingOutboundBytes(size, false, notifyWritability);
+        }
+
+        // recycle the entry
+        e.recycle();
+
+        return true;
+    }
+
+    private void removeEntry(Entry e) {
+        if (-- flushed == 0) {
+            // processed everything
+            flushedEntry = null;
+            if (e == tailEntry) {
+                tailEntry = null;
+                unflushedEntry = null;
+            }
+        } else {
+            flushedEntry = e.next;
+        }
+    }
+
+    @Override
+    public void removeBytes(long writtenBytes) {
+        Object msg = current();
+        while (writtenBytes > 0 || hasZeroReadable(msg)) {
+            if (msg instanceof Buffer) {
+                Buffer buf = (Buffer) msg;
+                final int readableBytes = buf.readableBytes();
+                if (readableBytes <= writtenBytes) {
+                    progress(readableBytes);
+                    writtenBytes -= readableBytes;
+                    remove();
+                } else { // readableBytes > writtenBytes
+                    buf.readSplit(Math.toIntExact(writtenBytes)).close();
+                    progress(writtenBytes);
+                    break;
+                }
+            } else if (msg instanceof ByteBufConvertible) {
+                final ByteBuf buf = ((ByteBufConvertible) msg).asByteBuf();
+                final int readerIndex = buf.readerIndex();
+                final int readableBytes = buf.writerIndex() - readerIndex;
+
+                if (readableBytes <= writtenBytes) {
+                    progress(readableBytes);
+                    writtenBytes -= readableBytes;
+                    remove();
+                } else { // readableBytes > writtenBytes
+                    buf.readerIndex(readerIndex + (int) writtenBytes);
+                    progress(writtenBytes);
+                    break;
+                }
+            } else {
+                break; // Don't know how to process this message. Might be null.
+            }
+            msg = current();
+        }
+    }
+
+    private static boolean hasZeroReadable(Object msg) {
+        if (msg instanceof Buffer) {
+            return ((Buffer) msg).readableBytes() == 0;
+        }
+        if (msg instanceof ByteBufConvertible) {
+            return ((ByteBufConvertible) msg).asByteBuf().readableBytes() == 0;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean isWritable() {
+        return unwritable == 0;
+    }
+
+    @Override
+    public boolean getUserDefinedWritability(int index) {
+        return (unwritable & writabilityMask(index)) == 0;
+    }
+
+    @Override
+    public void setUserDefinedWritability(int index, boolean writable) {
+        if (writable) {
+            setUserDefinedWritability(index);
+        } else {
+            clearUserDefinedWritability(index);
+        }
+    }
+
+    private void setUserDefinedWritability(int index) {
+        final int mask = ~writabilityMask(index);
+        for (;;) {
+            final int oldValue = unwritable;
+            final int newValue = oldValue & mask;
+            if (UNWRITABLE_UPDATER.compareAndSet(this, oldValue, newValue)) {
+                if (oldValue != 0 && newValue == 0) {
+                    fireChannelWritabilityChanged(true);
+                }
+                break;
+            }
+        }
+    }
+
+    private void clearUserDefinedWritability(int index) {
+        final int mask = writabilityMask(index);
+        for (;;) {
+            final int oldValue = unwritable;
+            final int newValue = oldValue | mask;
+            if (UNWRITABLE_UPDATER.compareAndSet(this, oldValue, newValue)) {
+                if (oldValue == 0 && newValue != 0) {
+                    fireChannelWritabilityChanged(true);
+                }
+                break;
+            }
+        }
+    }
+
+    private static int writabilityMask(int index) {
+        if (index < 1 || index > 31) {
+            throw new IllegalArgumentException("index: " + index + " (expected: 1~31)");
+        }
+        return 1 << index;
+    }
+
+    private void setWritable(boolean invokeLater) {
+        for (;;) {
+            final int oldValue = unwritable;
+            final int newValue = oldValue & ~1;
+            if (UNWRITABLE_UPDATER.compareAndSet(this, oldValue, newValue)) {
+                if (oldValue != 0 && newValue == 0) {
+                    fireChannelWritabilityChanged(invokeLater);
+                }
+                break;
+            }
+        }
+    }
+
+    private void setUnwritable(boolean invokeLater) {
+        for (;;) {
+            final int oldValue = unwritable;
+            final int newValue = oldValue | 1;
+            if (UNWRITABLE_UPDATER.compareAndSet(this, oldValue, newValue)) {
+                if (oldValue == 0) {
+                    fireChannelWritabilityChanged(invokeLater);
+                }
+                break;
+            }
+        }
+    }
+
+    private void fireChannelWritabilityChanged(boolean invokeLater) {
+        final ChannelPipeline pipeline = channel.pipeline();
+        if (invokeLater) {
+            Runnable task = fireChannelWritabilityChangedTask;
+            if (task == null) {
+                fireChannelWritabilityChangedTask = task = pipeline::fireChannelWritabilityChanged;
+            }
+            channel.executor().execute(task);
+        } else {
+            pipeline.fireChannelWritabilityChanged();
+        }
+    }
+
+    @Override
+    public int size() {
+        return flushed;
+    }
+
+    @Override
+    public boolean isEmpty() {
+        return flushed == 0;
+    }
+
+    void failFlushed(Throwable cause, boolean notify) {
+        // Make sure that this method does not reenter.  A listener added to the current promise can be notified by the
+        // current thread in the tryFailure() call of the loop below, and the listener can trigger another fail() call
+        // indirectly (usually by closing the channel.)
+        //
+        // See https://github.com/netty/netty/issues/1501
+        if (inFail) {
+            return;
+        }
+
+        try {
+            inFail = true;
+            for (;;) {
+                if (!remove0(cause, notify)) {
+                    break;
+                }
+            }
+        } finally {
+            inFail = false;
+        }
+    }
+
+    void close(final Throwable cause, final boolean allowChannelOpen) {
+        if (inFail) {
+            channel.executor().execute(() -> close(cause, allowChannelOpen));
+            return;
+        }
+
+        inFail = true;
+
+        if (!allowChannelOpen && channel.isOpen()) {
+            throw new IllegalStateException("close() must be invoked after the channel is closed.");
+        }
+
+        if (!isEmpty()) {
+            throw new IllegalStateException("close() must be invoked after all flushed writes are handled.");
+        }
+
+        // Release all unflushed messages.
+        try {
+            Entry e = unflushedEntry;
+            while (e != null) {
+                // Just decrease; do not trigger any events via decrementPendingOutboundBytes()
+                int size = e.pendingSize;
+                TOTAL_PENDING_SIZE_UPDATER.addAndGet(this, -size);
+
+                if (!e.cancelled) {
+                    SilentDispose.dispose(e.msg, logger);
+                    safeFail(e.promise, cause);
+                }
+                e = e.recycleAndGetNext();
+            }
+        } finally {
+            inFail = false;
+        }
+    }
+
+    void close(ClosedChannelException cause) {
+        close(cause, false);
+    }
+
+    private static void safeSuccess(Promise<Void> promise) {
+        PromiseNotificationUtil.trySuccess(promise, null, logger);
+    }
+
+    private static void safeFail(Promise<Void> promise, Throwable cause) {
+        PromiseNotificationUtil.tryFailure(promise, cause, logger);
+    }
+
+    @Override
+    public long totalPendingWriteBytes() {
+        return totalPendingSize;
+    }
+
+    @Override
+    public long bytesBeforeUnwritable() {
+        long bytes = channel.config().getWriteBufferHighWaterMark() - totalPendingSize;
+        // If bytes is negative we know we are not writable, but if bytes is non-negative we have to check writability.
+        // Note that totalPendingSize and isWritable() use different volatile variables that are not synchronized
+        // together. totalPendingSize will be updated before isWritable().
+        if (bytes > 0) {
+            return isWritable() ? bytes : 0;
+        }
+        return 0;
+    }
+
+    @Override
+    public long bytesBeforeWritable() {
+        long bytes = totalPendingSize - channel.config().getWriteBufferLowWaterMark();
+        // If bytes is negative we know we are writable, but if bytes is non-negative we have to check writability.
+        // Note that totalPendingSize and isWritable() use different volatile variables that are not synchronized
+        // together. totalPendingSize will be updated before isWritable().
+        if (bytes > 0) {
+            return isWritable() ? 0 : bytes;
+        }
+        return 0;
+    }
+
+    @Override
+    public void forEachFlushedMessage(MessageProcessor processor) throws Exception {
+        requireNonNull(processor, "processor");
+
+        Entry entry = flushedEntry;
+        if (entry == null) {
+            return;
+        }
+
+        do {
+            if (!entry.cancelled) {
+                if (!processor.processMessage(entry.msg)) {
+                    return;
+                }
+            }
+            entry = entry.next;
+        } while (isFlushedEntry(entry));
+    }
+
+    private boolean isFlushedEntry(Entry e) {
+        return e != null && e != unflushedEntry;
+    }
+
+    private static final class Entry {
+        private static final ObjectPool<Entry> RECYCLER = ObjectPool.newPool(Entry::new);
+
+        private final Handle<Entry> handle;
+        Entry next;
+        Object msg;
+        ByteBuffer[] bufs;
+        ByteBuffer buf;
+        Promise<Void> promise;
+        long progress;
+        long total;
+        int pendingSize;
+        int count = -1;
+        boolean cancelled;
+
+        private Entry(Handle<Entry> handle) {
+            this.handle = handle;
+        }
+
+        static Entry newInstance(Object msg, int size, long total, Promise<Void> promise) {
+            Entry entry = RECYCLER.get();
+            entry.msg = msg;
+            entry.pendingSize = size + CHANNEL_OUTBOUND_BUFFER_ENTRY_OVERHEAD;
+            entry.total = total;
+            entry.promise = promise;
+            return entry;
+        }
+
+        int cancel() {
+            if (!cancelled) {
+                cancelled = true;
+                int pSize = pendingSize;
+
+                // release message and replace with an empty buffer
+                SilentDispose.dispose(msg, logger);
+                msg = Unpooled.EMPTY_BUFFER;
+
+                pendingSize = 0;
+                total = 0;
+                progress = 0;
+                bufs = null;
+                buf = null;
+                return pSize;
+            }
+            return 0;
+        }
+
+        void recycle() {
+            next = null;
+            bufs = null;
+            buf = null;
+            msg = null;
+            promise = null;
+            progress = 0;
+            total = 0;
+            pendingSize = 0;
+            count = -1;
+            cancelled = false;
+            handle.recycle(this);
+        }
+
+        Entry recycleAndGetNext() {
+            Entry next = this.next;
+            recycle();
+            return next;
+        }
+    }
+}
